requirements.txt
appdirs==1.4.4
blessed==1.20.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cloudpickle==3.1.1
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
gpustat==1.1.1
gym==0.26.2
gym-notices==0.0.8
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lapx==0.5.11
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
names==0.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
pathtools==0.1.2
Pillow==10.1.0
protobuf==4.24.4
psutil==5.9.6
py-cpuinfo==9.0.0
pyaml==24.9.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scikit-optimize==0.10.2
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
torch==2.1.0
torchsummary==1.5.1
torchvision==0.16.0
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.1.34
urllib3==2.0.7
wandb==0.15.12
wcwidth==0.2.13
zipp==3.17.0
